Commit before deleting incorrect app.py and load_data.py files
