+++ tests/data/ehr/gp_clinical.txt
@@ -1,1 +1,7 @@
 eid	data_provider	event_dt	read_2	read_3	value1	value2	value3
+1	2	22/03/2014	j550.
+2	1	17/03/2016	42Z7.		12.800
+2	1	17/03/2016	42Z7.		12.800
+3	1	04/11/2013	426..		4.000
+3	1	04/11/2013	4266.		0.000
+3	1	04/11/2013	428..		28.700
