+++ tests/data/ehr/gp_registrations.txt
@@ -1,1 +1,5 @@
 eid	data_provider	reg_date	deduct_date
+1	2	02/02/1902
+2	1	12/06/2002	21/03/2014
+2	1	25/04/2014
+3	1	30/09/1993
