+++ ukbrest/common/pheno2sql.py
@@ -18,62 +18,8 @@ from ukbrest.common.utils.datagen import get_tmpdir
 from ukbrest.common.utils.constants import BGEN_SAMPLES_TABLE, ALL_EIDS_TABLE
 from ukbrest.config import logger, SQL_CHUNKSIZE_ENV
 from ukbrest.common.utils.misc import get_list
+from ukbrest.common.loadsql import LoadSQL
 from ukbrest.resources.exceptions import UkbRestSQLExecutionError, UkbRestProgramExecutionError
-
-class LoadSQL(DBAccess):
-    def __init__(self, db_uri, n_columns_per_table=sys.maxsize,
-                 loading_n_jobs=-1, tmpdir=tempfile.mkdtemp(prefix='ukbrest'),
-                 loading_chunksize=5000, sql_chunksize=None,
-                 delete_temp_csv=True):
-        """
-
-        :param db_uri:
-        :param n_columns_per_table:
-        :param loading_n_jobs:
-        :param tmpdir:
-        :param loading_chunksize:
-        :param sql_chunksize:
-        :param delete_temp_csv:
-        """
-        super(LoadSQL, self).__init__(db_uri)
-        self.n_columns_per_table = n_columns_per_table
-        logger.debug("n_columns_per_table set to {}".format(self.n_columns_per_table))
-        self.loading_n_jobs = loading_n_jobs
-        self.tmpdir = tmpdir
-        self.loading_chunksize = loading_chunksize
-
-        self.sql_chunksize = sql_chunksize
-        if self.sql_chunksize is None:
-            logger.warning('{} was not set, no chunksize for SQL queries, what can lead to '
-                           'memory problems.'.format(SQL_CHUNKSIZE_ENV))
-        self.delete_temp_csv = delete_temp_csv
-
-    def _run_psql(self, sql_statement, is_file=False):
-        current_env = os.environ.copy()
-        current_env['PGPASSWORD'] = self.db_pass
-
-        p = Popen(['psql', '-w', '-h', self.db_host, '-p', str(self.db_port),
-                   '-U', self.db_user, '-d', self.db_name,
-                   '-f' if is_file else '-c', sql_statement],
-                  stdout=PIPE, stderr=PIPE, env=current_env)
-
-        stdout_data, stderr_data = p.communicate()
-        stdout_data = stdout_data.decode('utf-8')
-        stderr_data = stderr_data.decode('utf-8')
-
-        if p.returncode != 0:
-            raise UkbRestSQLExecutionError(stdout_data + '\n' + stderr_data)
-        elif stderr_data is not None and 'ERROR:' in stderr_data:
-            raise UkbRestSQLExecutionError(stderr_data)
-
-    def _vacuum(self):
-        logger.info('Vacuuming')
-
-        with self._get_db_engine().connect().execution_options(isolation_level="AUTOCOMMIT") as conn:
-            conn.execute("""
-                vacuum analyze;
-            """)
-
 
 class Pheno2SQL(LoadSQL):
     _RE_COLUMN_NAME_PATTERN = '(?i)c[0-9a-z_]+_[0-9]+_[0-9]+'
@@ -115,37 +61,10 @@ class Pheno2SQL(LoadSQL):
 
         self.bgen_sample_file = bgen_sample_file
 
-        # parse_result = urlparse(self.db_uri)
-        # self.db_type = parse_result.scheme
-        #
-        # if self.db_type == 'sqlite':
-        #     logger.warning('sqlite does not support parallel loading')
-        #     self.db_file = self.db_uri.split(':///')[-1]
-        # elif self.db_type == 'postgresql':
-        #     self.db_host = parse_result.hostname
-        #     self.db_port = parse_result.port
-        #     self.db_name = parse_result.path.split('/')[-1]
-        #     self.db_user = parse_result.username
-        #     self.db_pass = parse_result.password
-
         self.table_prefix = table_prefix
-        # self.n_columns_per_table = n_columns_per_table
-        # logger.debug("n_columns_per_table set to {}".format(self.n_columns_per_table))
-        # self.loading_n_jobs = loading_n_jobs
-        # self.tmpdir = tmpdir
-        # self.loading_chunksize = loading_chunksize
-        #
-        # self.sql_chunksize = sql_chunksize
-        # if self.sql_chunksize is None:
-        #     logger.warning('{} was not set, no chunksize for SQL queries, what can lead to '
-        #                    'memory problems.'.format(SQL_CHUNKSIZE_ENV))
-
-        self._fields_dtypes = {}
 
         # this is a temporary variable that holds information about loading
         self._loading_tmp = {}
-
-        self.table_list = set()
 
         self.csv_files_encoding_file = 'encodings.txt'
         self.csv_files_encoding = 'utf-8'
@@ -429,8 +348,6 @@ class Pheno2SQL(LoadSQL):
             for column_names_idx, column_names in self._loading_tmp['chunked_column_names']
         )
 
-        self.table_list.update(table_name for table_name, file_path in self.table_csvs)
-
     def _load_single_csv(self, table_name, file_path):
         logger.info('{} -> {}'.format(file_path, table_name))
 
@@ -486,37 +403,6 @@ class Pheno2SQL(LoadSQL):
             for table_name, file_path in self.table_csvs:
                 self._load_single_csv(table_name, file_path)
 
-    def _load_all_eids(self):
-        logger.info('Loading all eids into table {}'.format(ALL_EIDS_TABLE))
-
-        create_table(ALL_EIDS_TABLE,
-            columns=[
-                'eid bigint NOT NULL',
-            ],
-            constraints=[
-                'pk_{} PRIMARY KEY (eid)'.format(ALL_EIDS_TABLE)
-            ],
-            db_engine=self._get_db_engine()
-         )
-
-        select_eid_sql = ' UNION DISTINCT '.join(
-            'select eid from {}'.format(table_name)
-            for table_name in self.table_list
-        )
-
-        insert_eids_sql = """
-            insert into {all_eids_table} (eid)
-            (
-                {sql_eids}
-            )
-        """.format(
-            all_eids_table=ALL_EIDS_TABLE,
-            sql_eids=select_eid_sql
-        )
-
-        with self._get_db_engine().connect() as con:
-            con.execute(insert_eids_sql)
-
     def _load_bgen_samples(self):
         if self.bgen_sample_file is None or not os.path.isfile(self.bgen_sample_file):
             logger.warning('BGEN sample file not set or does not exist: {}'.format(self.bgen_sample_file))
@@ -541,24 +427,6 @@ class Pheno2SQL(LoadSQL):
         samples_data.rename(columns={'ID_1': 'eid'}, inplace=True)
 
         samples_data.to_sql(BGEN_SAMPLES_TABLE, self._get_db_engine(), if_exists='append')
-
-    # def _run_psql(self, sql_statement, is_file=False):
-    #     current_env = os.environ.copy()
-    #     current_env['PGPASSWORD'] = self.db_pass
-    #
-    #     p = Popen(['psql', '-w', '-h', self.db_host, '-p', str(self.db_port),
-    #                '-U', self.db_user, '-d', self.db_name,
-    #                '-f' if is_file else '-c', sql_statement],
-    #               stdout=PIPE, stderr=PIPE, env=current_env)
-    #
-    #     stdout_data, stderr_data = p.communicate()
-    #     stdout_data = stdout_data.decode('utf-8')
-    #     stderr_data = stderr_data.decode('utf-8')
-    #
-    #     if p.returncode != 0:
-    #         raise UkbRestSQLExecutionError(stdout_data + '\n' + stderr_data)
-    #     elif stderr_data is not None and 'ERROR:' in stderr_data:
-    #         raise UkbRestSQLExecutionError(stderr_data)
 
     def _load_events(self):
         if self.db_type == 'sqlite':
@@ -630,14 +498,6 @@ class Pheno2SQL(LoadSQL):
         create_indexes('events', ('eid', 'field_id', 'instance', 'event', ('field_id', 'event')),
                        db_engine=self._get_db_engine())
 
-    # def _vacuum(self):
-    #     logger.info('Vacuuming')
-    #
-    #     with self._get_db_engine().connect().execution_options(isolation_level="AUTOCOMMIT") as conn:
-    #         conn.execute("""
-    #             vacuum analyze;
-    #         """)
-
     def load_data(self, vacuum=False):
         """
         Load all CSV files specified into the database configured.
@@ -683,658 +543,3 @@ class Pheno2SQL(LoadSQL):
         self.init_field_dtypes()
 
         logger.info('Initialization finished!')
-
-    def _create_joins(self, tables, join_type='inner join'):
-        if len(tables) == 0:
-            return ""
-
-        if len(tables) == 1:
-            return tables[0]
-
-        return tables[0] + ' ' + ' '.join([
-            '{join_type} {table} using (eid) '.format(join_type=join_type, table=t) for t in tables[1:]
-        ])
-
-    def _get_needed_tables(self, all_columns):
-        if len(all_columns) == 0:
-            return []
-
-        all_columns_quoted = ["'{}'".format(x.replace("'", "''")) for x in all_columns]
-
-        tables_needed_df = pd.read_sql(
-            'select distinct table_name '
-            'from fields '
-            'where column_name in (' + ','.join(all_columns_quoted) + ')',
-        self._get_db_engine()).loc[:, 'table_name'].tolist()
-
-        if len(tables_needed_df) == 0:
-            return []
-
-        return tables_needed_df
-
-    def get_field_dtype(self, field=None):
-        """Returns the type of the field. If field is None, then it just loads all fields types"""
-
-        if field in self._fields_dtypes:
-            return self._fields_dtypes[field]
-
-        # initialize dbtypes for all fields
-        field_type = pd.read_sql(
-            'select distinct column_name, type '
-            'from fields',
-        self._get_db_engine())
-
-        for row in field_type.itertuples():
-            self._fields_dtypes[row.column_name] = row.type
-
-        return self._fields_dtypes[field] if field in self._fields_dtypes else None
-
-    def _get_fields_from_reg_exp(self, ecolumns):
-        if ecolumns is None:
-            return []
-
-        where_st = ["column_name ~ '{}'".format(ecol) for ecol in ecolumns]
-        select_st = """
-            select distinct column_name
-            from fields
-            where {}
-            order by column_name
-        """.format(' or '.join(where_st))
-
-        return pd.read_sql(select_st, self._get_db_engine()).loc[:, 'column_name'].tolist()
-
-    def _get_fields_from_statements(self, statement):
-        """This method gets all fields mentioned in the statements."""
-        columns_fields = []
-        if statement is not None:
-            columns_fields = list(set([x for col in statement for x in re.findall(Pheno2SQL.RE_COLUMN_NAME, col)]))
-
-        return columns_fields
-
-    def _get_integer_fields(self, columns):
-        """This method returns a list of fields (either its column specification, like c64_0_0 or its rename like
-        myfield) that are of type integer."""
-        int_columns = []
-
-        for col in columns:
-            if col == 'eid':
-                continue
-
-            match = re.search(Pheno2SQL.RE_FULL_COLUMN_NAME_RENAME, col)
-
-            if match is None:
-                continue
-
-            col_field = match.group('field')
-
-            if self.get_field_dtype(col_field) != 'Integer':
-                continue
-
-            # select rename first, if not specified select field column
-            col_rename = next((grp_val for grp_val in (match.group('rename'), match.group('field')) if grp_val is not None))
-            int_columns.append(col_rename)
-
-        return int_columns
-
-    def _get_filterings(self, filter_statements):
-        return ' AND '.join('({})'.format(afilter) for afilter in filter_statements)
-
-    def _query_generic(self, sql_query, order_by_dict=None, results_transformator=None):
-        final_sql_query = sql_query
-
-        if order_by_dict is not None:
-            outer_sql = """
-                select {data_fields}
-                from {order_by} s left outer join (
-                    {base_sql}
-                ) u
-                using (eid)
-                order by s.index asc
-            """.format(
-                order_by=order_by_dict['table'],
-                base_sql=sql_query,
-                data_fields=order_by_dict['columns_select']
-            )
-
-            final_sql_query = outer_sql
-
-        logger.debug(final_sql_query)
-
-        try:
-            results_iterator = pd.read_sql(
-                final_sql_query, self._get_db_engine(), index_col='eid', chunksize=self.sql_chunksize
-            )
-        except ProgrammingError as e:
-            raise UkbRestSQLExecutionError(str(e))
-
-        if self.sql_chunksize is None:
-            results_iterator = iter([results_iterator])
-
-        for chunk in results_iterator:
-            if results_transformator is not None:
-                chunk = results_transformator(chunk)
-
-            yield chunk
-
-    def _get_query_sql(self, columns=None, ecolumns=None, filterings=None):
-        # select needed tables to join
-        columns_fields = self._get_fields_from_statements(columns)
-        reg_exp_columns_fields = self._get_fields_from_reg_exp(ecolumns)
-        filterings_columns_fields = self._get_fields_from_statements(filterings)
-
-        tables_needed_df = self._get_needed_tables(columns_fields + reg_exp_columns_fields + filterings_columns_fields)
-
-        all_columns = ['eid'] + (columns if columns is not None else []) + reg_exp_columns_fields
-
-        base_sql = """
-            select {data_fields}
-            {from_clause}
-            {where_statements}
-        """
-
-        tables_join_sql = self._create_joins(tables_needed_df, join_type='full outer join')
-
-        if tables_join_sql:
-            from_clause_sql = f'from {tables_join_sql}'
-        else:
-            from_clause_sql = 'from all_eids'
-
-
-        return base_sql.format(
-            data_fields=','.join(all_columns),
-            from_clause=from_clause_sql,
-            where_statements=((' where ' + self._get_filterings(filterings)) if filterings is not None else ''),
-        )
-
-    def query(self, columns=None, ecolumns=None, filterings=None, order_by_table=None):
-        reg_exp_columns_fields = self._get_fields_from_reg_exp(ecolumns)
-        all_columns = ['eid'] + (columns if columns is not None else []) + reg_exp_columns_fields
-
-        order_by_dict = None
-        if order_by_table is not None:
-            order_by_dict = {
-                'table': order_by_table,
-                'columns_select': ','.join(all_columns),
-            }
-
-        int_columns = self._get_integer_fields(all_columns)
-
-        final_sql_query = self._get_query_sql(columns, ecolumns, filterings)
-
-        def format_integer_columns(chunk):
-            for col in int_columns:
-                chunk[col] = chunk[col].map(lambda x: np.nan if pd.isnull(x) else '{:1.0f}'.format(x))
-
-            return chunk
-
-        return self._query_generic(
-            final_sql_query,
-            results_transformator=format_integer_columns,
-            order_by_dict=order_by_dict
-        )
-
-    def query_yaml_simple_data(self, yaml_file, section, order_by_table=None):
-        section_data = yaml_file[section]
-
-        include_only_stmts = None
-        if 'samples_filters' in yaml_file:
-            include_only_stmts = yaml_file['samples_filters']
-
-        section_field_statements = ['({}) as {}'.format(v, x) for x, v in section_data.items()]
-
-        for chunk in self.query(section_field_statements, filterings=include_only_stmts, order_by_table=order_by_table):
-            # chunk = chunk.rename(columns={v:k for x in section_data.items()})
-            yield chunk
-
-    def query_yaml_data(self, yaml_file, section, order_by_table=None):
-        all_columns = []
-        all_columns_sql_queries = []
-
-        where_st = ''
-        where_fields = []
-        if 'samples_filters' in yaml_file:
-            where_st = self._get_filterings(yaml_file['samples_filters'])
-            where_fields = self._get_fields_from_statements([where_st])
-
-        for column, column_dict in yaml_file[section].items():
-            all_columns.append(column)
-
-            subqueries = []
-
-            if isinstance(column_dict, dict):
-                for df, df_cods in column_dict.items():
-                    if df == 'sql':
-                        for cat_code, cat_condition in df_cods.items():
-                            # TODO: check for repeated category codes
-                            needed_tables = self._get_needed_tables(
-                                self._get_fields_from_statements([cat_condition]) + where_fields
-                            )
-
-                            sql_code = """
-                                select eid, {cat_code} as {column_name}
-                                from {cases_joins}
-                                {where_st}
-                            """.format(
-                                    cat_code=cat_code,
-                                    column_name=column,
-                                    cases_joins=self._create_joins(needed_tables + [ALL_EIDS_TABLE]),
-                                    where_st='where ({}) {}'.format(cat_condition, (' AND ({})'.format(where_st) if where_st else ''))
-                            )
-
-                            subqueries.append(sql_code)
-
-                    elif df == 'case_control':
-                        cases_conditions = [
-                            '(field_id = {} and event in ({}))'.format(
-                                field_id, ', '.join("'{}'".format(cod) for cod in get_list(field_cond['coding']))
-                            ) for field_id, field_cond in df_cods.items()
-                        ]
-
-                        sql_cases = """
-                            select distinct eid
-                            from events
-                            where {conditions}
-                        """.format(conditions=' OR '.join(cases_conditions))
-
-                        sql_cases_code = """
-                                select eid, 1 as {column_name}
-                                from {cases_joins}
-                                {where_st}
-                        """.format(
-                            column_name=column,
-                            cases_joins=self._create_joins([
-                                '({}) ev'.format(sql_cases)
-                            ] + self._get_needed_tables(where_fields)),
-                            where_st='where ' + (where_st if where_st else '1=1')
-                        )
-
-                        subqueries.append(sql_cases_code)
-
-                        # controls
-                        sql_controls_code = """
-                            select aet.eid, 0 as {column_name}
-                            from {controls_joins}
-                            {where_st}
-                            and aet.eid not in (
-                                {sql_cases}
-                            )
-                        """.format(
-                            column_name=column,
-                            controls_joins=self._create_joins([
-                                '{} aet'.format(ALL_EIDS_TABLE),
-                            ] + self._get_needed_tables(where_fields)),
-                            where_st='where ' + (where_st if where_st else '1=1'),
-                            sql_cases=sql_cases,
-                        )
-
-                        subqueries.append(sql_controls_code)
-
-                    else:
-                        raise Exception('Invalid selector type')
-
-            elif isinstance(column_dict, str):
-                final_sql = self._get_query_sql(
-                    columns=['({}) as {}'.format(column_dict, column)],
-                    filterings=yaml_file['samples_filters'] if 'samples_filters' in yaml_file else None,
-                )
-
-                subqueries.append(final_sql)
-            else:
-                raise Exception('Invalid query type')
-
-            column_sql_query = ' union distinct '.join(subqueries)
-
-            all_columns_sql_queries.append(column_sql_query)
-
-        order_by_dict = None
-        if order_by_table is not None:
-            order_by_dict = {
-                'table': order_by_table,
-                'columns_select':
-                    's.eid as eid, ' +
-                    ', '.join('{column_name}::text'.format(column_name=column) for column in all_columns),
-            }
-
-        final_sql_query = """
-            select eid, {columns_names}
-            from {inner_queries}
-        """.format(
-            columns_names=', '.join('{}::text'.format(column) for column in all_columns),
-            inner_queries=self._create_joins(
-                ['({}) iq{}'.format(iq, iq_idx) for iq_idx, iq in enumerate(all_columns_sql_queries)],
-                join_type='full outer join'
-            ),
-        )
-
-        return self._query_generic(
-            final_sql_query,
-            order_by_dict=order_by_dict
-        )
-
-    def query_yaml(self, yaml_file, section, order_by_table=None):
-        if section.startswith('simple_'):
-            return self.query_yaml_simple_data(yaml_file, section, order_by_table)
-        else:
-            return self.query_yaml_data(yaml_file, section, order_by_table)
-
-
-class EHR2SQL(LoadSQL):
-    K_CLINICAL = "gp_clinical"
-    K_SCRIPTS = "gp_scripts"
-    K_REGISTRATIONS = "gp_registrations"
-    DD_GP = {K_CLINICAL:"gp_clinical.txt",
-             K_SCRIPTS:"gp_scripts.txt",
-             K_REGISTRATIONS:"gp_registrations.txt"}
-    K_HESIN = "HESIN"
-    K_DIAG = "HESIN_DIAG"
-    DD_HESIN = {K_HESIN:"hesin.txt",
-                K_DIAG:"hesin_diag.txt"}
-
-    def __init__(self, db_uri, primary_care_dir, hospital_inpatient_dir,
-                 n_columns_per_table=sys.maxsize,
-                 loading_n_jobs=-1, tmpdir=tempfile.mkdtemp(prefix='ukbrest'),
-                 loading_chunksize=5000, sql_chunksize=None,
-                 delete_temp_csv=True):
-
-        super(EHR2SQL, self).__init__(db_uri,
-                                      n_columns_per_table=n_columns_per_table,
-                                      loading_n_jobs=loading_n_jobs,
-                                      tmpdir=tmpdir,
-                                      loading_chunksize=loading_chunksize,
-                                      sql_chunksize=sql_chunksize,
-                                      delete_temp_csv=delete_temp_csv)
-        if (hospital_inpatient_dir is None and
-            primary_care_dir is None):
-            raise ValueError("Neither hospital inpatient nor primary care "
-                             "directories were specified.")
-        self.primary_care_dir = primary_care_dir
-        if self.primary_care_dir is not None:
-            self.gp_file_dd = {}
-            for k,v in EHR2SQL.DD_GP.items():
-                fp = os.path.join(self.primary_care_dir, v)
-                if os.path.isfile(fp):
-                    self.gp_file_dd[k] = fp
-                else:
-                    raise ValueError(v + " out of place")
-        else:
-            self.gp_file_dd = None
-        self.hospital_inpatient_dir = hospital_inpatient_dir
-        if self.hospital_inpatient_dir is not None:
-            self.hesin_file_dd = {}
-            for k,v in EHR2SQL.DD_HESIN.items():
-                fp = os.path.join(self.hospital_inpatient_dir, v)
-                if os.path.isfile(fp):
-                    self.hesin_file_dd[k] = fp
-                else:
-                    raise ValueError(v + " out of place")
-        else: 
-            self.hesin_file_dd = None
-
-    def load_data(self, vacuum):
-        logger.info("Loading EHR into database")
-        self._load_primary_care_data()
-        self._load_hospital_inpatient_data()
-        # self._create_constraints()
-
-        if vacuum:
-            self._vacuum()
-
-    def _load_primary_care_data(self):
-        
-        db_engine = self._get_db_engine()
-
-        # PG REGISTRATIONS
-        self._create_gp_registrations_table()
-        # gp_registrations_df = self._load_gp_registrations_df()
-        gp_registrations_df = self._load_ehr_df(self.gp_file_dd[EHR2SQL.K_REGISTRATIONS],
-                                                ['eid', 'reg_date'],
-                                                day_date_cols=['reg_date', 'deduct_date'])
-        gp_registrations_df.to_sql(EHR2SQL.K_REGISTRATIONS, db_engine,
-                                   if_exists='append', index=False)
-
-        # PG CLINICAL
-        self._create_gp_clinical_table()
-        # gp_clinical_df = self._load_gp_clinical_df()
-        gp_clinical_df = self._load_ehr_df(self.gp_file_dd[EHR2SQL.K_CLINICAL],
-                                           ['eid', 'event_dt', 'read_key'],
-                                           day_date_cols=['event_dt'],
-                                           accumulate_col_dict={'read_key':['read_2', 'read_3']})
-        gp_clinical_df.to_sql(EHR2SQL.K_CLINICAL, db_engine, if_exists='append',
-                              index=False)
-
-        # PG SCRIPTS
-        self._create_gp_scripts_table()
-        # gp_scripts_df = self._load_gp_scripts_df()
-        gp_scripts_df = self._load_ehr_df(self.gp_file_dd[EHR2SQL.K_SCRIPTS],
-                                          ['eid', 'issue_date', 'read_key'],
-                                          day_date_cols=['issue_date'],
-                                          accumulate_col_dict={'read_key':['bnf_code', 'dmd_code', 'read_2']},
-                                          dtype_handling={'bnf_code':str})
-        gp_scripts_df.to_sql(EHR2SQL.K_SCRIPTS, db_engine, if_exists='append',
-                             index=False)
-
-    # def _load_gp_scripts_df(self):
-    #     fp = self.gp_file_dd[EHR2SQL.K_SCRIPTS]
-    #     pk_cols = ['eid', 'issue_date', 'read_key']
-    #     date_col = 'issue_date'
-    #     logger.info("Loading table: {}".format(fp))
-    #     gp_df = pd.read_table(fp, encoding='latin1', dtype={'bnf_code': str})
-    #     gp_df['read_key'] = gp_df['bnf_code'].fillna(gp_df['dmd_code'])
-    #     gp_df['read_key'] = gp_df['read_key'].fillna(gp_df['read_2'])
-    #     o_len = gp_df.shape[0]
-    #     gp_df = gp_df.drop_duplicates(pk_cols)
-    #     new_len = gp_df.shape[0]
-    #     logger.warning("Dropped {} entries from table with duplicated cols: {}".format(o_len - new_len,
-    #                                                                                    pk_cols))
-    #     gp_df[date_col] = pd.to_datetime(gp_df[date_col], dayfirst=True)
-    #     return gp_df
-
-    def _create_gp_scripts_table(self):
-        create_table(EHR2SQL.K_SCRIPTS, columns = [
-                            'eid bigint NOT NULL',
-                            'data_provider int NOT NULL',
-                            'issue_date date NOT NULL',
-                            'read_key text NOT NULL',
-                            'read_2 text',
-                            'bnf_code text',
-                            'dmd_code text',
-                            'drug_name text',
-                            'quantity text'
-                        ],
-                     constraints=[
-                         'pk_{} PRIMARY KEY (eid, issue_date, read_key)'.format(EHR2SQL.K_SCRIPTS)
-                        ],
-                        db_engine=self._get_db_engine())
-
-    def _create_gp_clinical_table(self):
-        create_table(EHR2SQL.K_CLINICAL, columns=[
-            'eid bigint NOT NULL',
-            'data_provider int NOT NULL',
-            'event_dt date NOT NULL',
-            'read_key text NOT NULL',
-            'read_2 text',
-            'read_3 text',
-            'value1 text',
-            'value2 text',
-            'value3 text',
-        ],
-                     constraints=[
-                         'pk_{} PRIMARY KEY (eid, event_dt, read_key)'.format(EHR2SQL.K_CLINICAL)
-                     ],
-                     db_engine=self._get_db_engine())
-
-    # def _load_gp_clinical_df(self):
-    #     fp = self.gp_file_dd[EHR2SQL.K_CLINICAL]
-    #     date_col = 'event_dt'
-    #     logger.info("Loading table: {}".format(fp))
-    #     gp_clinical_df = pd.read_table(fp,
-    #                           encoding='latin1', dtype={'value1': str,
-    #                                                     'value2': str,
-    #                                                     'value3': str})
-    #     gp_clinical_df['read_key'] = gp_clinical_df['read_2'].fillna(gp_clinical_df['read_3'])
-    #     o_len = gp_clinical_df.shape[0]
-    #     gp_clinical_df = gp_clinical_df.drop_duplicates(['eid', 'event_dt', 'read_key'])
-    #     new_len = gp_clinical_df.shape[0]
-    #     logger.warning("Dropped {} entries from table with duplicated cols: {}".format(o_len - new_len,
-    #                                                                     ['eid', 'event_dt', 'read_key']))
-    #     gp_clinical_df[date_col] = pd.to_datetime(gp_clinical_df[date_col], dayfirst=True)
-    #     return gp_clinical_df
-
-    def _create_gp_registrations_table(self):
-        create_table(EHR2SQL.K_REGISTRATIONS, columns=[
-            'eid bigint NOT NULL',
-            'data_provider int NOT NULL',
-            'reg_date date NOT NULL',
-            'deduct_date date',
-        ],
-                     constraints=[
-                         'pk_{} PRIMARY KEY (eid, reg_date)'.format(EHR2SQL.K_REGISTRATIONS)
-                     ],
-                     db_engine=self._get_db_engine())
-
-    # def _load_gp_registrations_df(self):
-    #     fp = self.gp_file_dd[EHR2SQL.K_REGISTRATIONS]
-    #     logger.info("Loading table: {}".format(fp))
-    #     df = pd.read_table(fp, encoding='latin1')
-    #     df['reg_date'] = pd.to_datetime(df['reg_date'], dayfirst=True)
-    #     df['deduct_date'] = pd.to_datetime(df['deduct_date'], dayfirst=True)
-    #     o_len = df.shape[0]
-    #     df = df.drop_duplicates(['eid', 'reg_date'])
-    #     n_len = df.shape[0]
-    #     logger.warning("Dropped {} entries from table with duplicated cols: {}".format(o_len - n_len,
-    #                                                                                    ['eid', 'reg_dt']))
-    #     df = df.dropna(axis=0, subset=['eid', 'reg_date'])
-    #     return df
-
-    def _load_hospital_inpatient_data(self):
-        db_engine = self._get_db_engine()
-
-        # HESIN
-        self._create_hesin_table()
-        hesin_df = self._load_ehr_df(self.hesin_file_dd[EHR2SQL.K_HESIN],
-                                     ['eid', 'ins_index'],
-                                     ymd_date_cols=['epistart', 'epiend',
-                                                    'elecdate', 'admidate'])
-        hesin_df.to_sql(EHR2SQL.K_HESIN, db_engine, if_exists='append',
-                              index=False)
-
-        # HESIN_DIAG
-        self._create_hesin_diag_table()
-        diag_df = self._load_ehr_df(self.hesin_file_dd[EHR2SQL.K_DIAG],
-                                    ['eid', 'ins_index', 'arr_index'])
-        diag_df.to_sql(EHR2SQL.K_DIAG, db_engine, if_exists='append',
-                       index=False)
-
-
-
-    def _create_hesin_table(self):
-        create_table(EHR2SQL.K_HESIN,
-                     ['eid bigint NOT NULL',
-                      'ins_index bigint NOT NULL',
-                      'dsource text NOT NULL',
-                      'source text NOT NULL',
-                      'epistart date',
-                      'epiend date',
-                      'epidur bigint',
-                      'bedyear int',
-                      'epistat int',
-                      'epitype int',
-                      'epiorder int',
-                      'spell_index int',
-                      'spell_seq int',
-                      'spelbgin int',
-                      'spelend int',
-                      'speldur bigint',
-                      'pctcode text',
-                      'gpprpct text',
-                      'category int',
-                      'elecdate date',
-                      'admidate date',
-                      'admimeth_uni int',
-                      'admimeth int',
-                      'admisorc_uni int',
-                      'asmisorc int',
-                      'firstreg int',
-                      'classpat_uni int',
-                      'classpat int',
-                      'intmanag_uni int',
-                      'intmanag int',
-                      'mainspef_uni int',
-                      'mainspef text',
-                      'tretspef_uni int',
-                      'operstat int',
-                      'disdate date',
-                      'dismeth_uni int',
-                      'dismeth int',
-                      'didest_uni int',
-                      'didest int'],
-                     db_engine=self._get_db_engine(), 
-                     constraints=['pk_{} PRIMARY KEY (eid, ins_index)'.format(EHR2SQL.K_HESIN)])
-
-    def _create_hesin_diag_table(self):
-        create_table(EHR2SQL.K_DIAG, columns=[
-                                    'eid bigint NOT NULL',
-                                    'ins_index bigint NOT NULL',
-                                    'arr_index bigint NOT NULL',
-                                    'level int',
-                                    'diag_icd9 text',
-                                    'diag_icd9_nb text',
-                                    'diag_icd10 text',
-                                    'diag_icd10_nb text'],
-                     db_engine=self._get_db_engine(),
-                     constraints=['pk_{} PRIMARY KEY (eid, ins_index, arr_index)'.format(EHR2SQL.K_DIAG)])
-
-    @staticmethod
-    def _load_ehr_df(fp, pk_cols, day_date_cols=None, month_date_cols=None,
-                     ymd_date_cols=None, encoding=None,
-                     accumulate_col_dict=None, extra_not_null = None,
-                     dtype_handling=None):
-        """
-
-        :param fp: filepath to EHR text file
-        :param pk_cols: columns treated as primary keys. Uniqueness and NonNull
-                      are enforced.
-        :param day_date_cols: Date columns to be treated with day first
-        :param month_date_cols: Date columns to be treated with month first
-        :param encoding: Text file encoding
-        :param accumulate_col_dict: {resulting col name: [ first_option,
-                                                            second_option,
-                                                            ...]}
-        :param extra_not_null: list of other Not Null columns
-        :return: pandas DataFrame
-        """
-        logger.info("Loading table {}".format(fp))
-        df = pd.read_table(fp, encoding=encoding, dtype=dtype_handling)
-        if day_date_cols is not None:
-            for c in day_date_cols:
-                df[c] = pd.to_datetime(df[c], dayfirst=True)
-        if month_date_cols is not None:
-            for c in month_date_cols:
-                df[c] = pd.to_datetime(df[c], monthfirst=True)
-        if ymd_date_cols is not None:
-            for c in ymd_date_cols:
-                df[c] = pd.to_datetime(df[c], format="%Y%m%d")
-
-        if accumulate_col_dict is not None:
-            for col, fill_cols in accumulate_col_dict.items():
-                df[col] = np.nan
-                for i in fill_cols:
-                    df[col] = df[col].fillna(df[i])
-
-        l_0 = df.shape[0]
-        df = df.drop_duplicates(pk_cols)
-        l_1 = df.shape[0]
-        if l_0 != l_1:
-            logger.warning("Dropped {} rows due to uniqueness constraint".format(l_0 - l_1))
-        if extra_not_null is not None:
-            pk_cols.extend(extra_not_null)
-        df = df.dropna(subset=pk_cols)
-        l_2 = df.shape[0]
-        if l_1 != l_2:
-            logger.warning("Dropped {} rows due to nonnull constraint".format(l_1 - l_2))
-
-        return df
-
-
-
-
