+++ ukbrest/common/utils/constants.py
@@ -1,3 +1,4 @@
 ALL_EIDS_TABLE='all_eids'
 WITHDRAWALS_TABLE='withdrawals'
 BGEN_SAMPLES_TABLE='bgen_samples'
+EXCLUDE_FROM_ALL_EIDS={'all_eids', 'fields'}
