+++ ukbrest/load_data.py
@@ -2,17 +2,19 @@ import os
 import argparse
 
 from ukbrest.common.pheno2sql import Pheno2SQL
+from ukbrest.common.ehr2sql import EHR2SQL
 from ukbrest.common.postloader import Postloader
 from ukbrest import config
 from ukbrest.common.utils.misc import update_parameters_from_args, parameter_empty
 
 from ukbrest.resources.error_handling import handle_errors
 
-parser = argparse.ArgumentParser()
+parser = argparse.ArgumentParser("Loads data into the SQL database")
 parser.add_argument('--load-sql', action='store_true')
-parser.add_argument('--load-withdrawals', action='store_true')
-parser.add_argument('--load-codings', action='store_true')
+parser.add_argument('--load-withdrawals', action='store_true', help="Loads withdrawals and then exits. Should be used with --withdrawals-dir")
+parser.add_argument('--load-codings', action='store_true', help="Loads codings and then exits. Should be used with --codings-dir.")
 parser.add_argument('--load-samples-data', action='store_true')
+parser.add_argument('--load-ehr', action='store_true', help="Load record-level EHR data")
 parser.add_argument('--identifier-columns', type=str, nargs='+', help='Format file1.txt:column1 file2.txt:column2 ...')
 parser.add_argument('--skip-columns', type=str, nargs='+', help='Format file1.txt:column1 file2.txt:column2 ...')
 parser.add_argument('--separators', type=str, nargs='+', help='Format file1.txt:column1 file2.txt:column2 ...')
@@ -20,14 +22,24 @@ parser.add_argument('--separators', type=str, nargs='+', help='Format file1.txt:
 
 @handle_errors
 def load_withdrawals(args):
-    pl = Postloader(**config.get_postloader_parameters())
-    pl.load_withdrawals(**config.get_postloader_withdrawals_parameters())
+    pl_params = config.get_postloader_parameters()
+    pl_params = update_parameters_from_args(pl_params, args)
+    pl = Postloader(**pl_params)
+
+    pl_wd_params = config.get_postloader_withdrawals_parameters()
+    pl_wd_params = update_parameters_from_args(pl_wd_params, args)
+    pl.load_withdrawals(**pl_wd_params)
 
 
 @handle_errors
 def load_codings(args):
-    pl = Postloader(**config.get_postloader_parameters())
-    pl.load_codings(**config.get_postloader_codings_parameters())
+    pl_params = config.get_postloader_parameters()
+    pl_params = update_parameters_from_args(pl_params, args)
+    pl = Postloader(**pl_params)
+
+    pl_coding_params = config.get_postloader_codings_parameters()
+    pl_coding_params = update_parameters_from_args(pl_coding_params, args)
+    pl.load_codings(**pl_coding_params)
 
 
 @handle_errors
@@ -71,7 +83,9 @@ def load_data(args):
 
     p2sql = Pheno2SQL(**pheno2sql_parameters)
 
-    p2sql.load_data()
+    load_parameters = config.get_pheno2sql_load_parameters()
+
+    p2sql.load_data(**load_parameters)
 
 
 @handle_errors
@@ -86,6 +100,20 @@ def load_sql():
 
     p2sql.load_sql('/opt/utils/sql/functions.sql')
 
+@handle_errors
+def load_ehr(args):
+    # pheno2sql_parameters = config.get_pheno2sql_parameters()
+    # pheno2sql_parameters = update_parameters_from_args(pheno2sql_parameters, args)
+
+    ehr2sql_parameters = config.get_ehr2sql_parameters()
+    ehr2sql_parameters = update_parameters_from_args(ehr2sql_parameters, args)
+
+    ehr2sql = EHR2SQL(**ehr2sql_parameters)
+
+    load_parameters = config.get_pheno2sql_load_parameters()
+
+    ehr2sql.load_data(**load_parameters)
+
 
 if __name__ == '__main__':
     parser = config.get_argparse_arguments(parser)
@@ -99,6 +127,9 @@ if __name__ == '__main__':
 
     elif args.load_samples_data:
         load_samples_data(args)
+
+    elif args.load_ehr:
+        load_ehr(args)
 
     elif args.load_sql:
         load_sql()
